+++ resources/views/admin/payments/index.blade.php
@@ -58,6 +58,7 @@
     {{ $payments->links() }}
 
     <div class="mt-4">
+        <a href="{{ route('admin.payments.renewals') }}" class="btn btn-success">{{ __('Membership renewals') }}</a>
         <a href="{{ route('admin.payments.reconciliation') }}" class="btn btn-primary">{{ __('Bank Reconciliation') }}</a>
     </div>
 </x-admin-layout>