+++ routes/admin.php
@@ -24,6 +24,7 @@ use App\Http\Controllers\Admin\MemberController;
 use App\Http\Controllers\Admin\MemberExportController;
 use App\Http\Controllers\Admin\NewsletterController;
 use App\Http\Controllers\Admin\PaymentController;
+use App\Http\Controllers\Admin\RenewalController;
 use App\Http\Controllers\Admin\RolePermissionController;
 use App\Http\Controllers\Admin\SeasonController;
 use App\Http\Controllers\Admin\SettingsController;
@@ -206,6 +207,8 @@ Route::get('/guide/{section}', [GuideController::class, 'show'])->name('guide.sh
 
 // Payments
 Route::get('/payments', [PaymentController::class, 'index'])->name('payments.index');
+Route::get('/payments/renewals', [RenewalController::class, 'index'])->name('payments.renewals');
+Route::post('/payments/renewals/{user}/received', [RenewalController::class, 'received'])->name('payments.renewals.received');
 Route::get('/payments/components', [PaymentController::class, 'components'])->name('payments.components');
 Route::post('/payments/components', [PaymentController::class, 'storeComponent'])->name('payments.component.store');
 Route::match(['put', 'patch'], '/payments/components/{component}', [PaymentController::class, 'updateComponent'])->name('payments.component.update');